Update DayTripGenerator.js
Pulled it out again to retest and found some more bugs.

After user types Yes, it finishes it but then re-runs the prompt and gives the user another random selection.

i also found that if you click no, and enter 1-4 it will not specifically change that item.
working on this...